Filtering UI & logic moved out of PeopleTable. . PeopleTable & PeopleList should be mostly dumb UI display components.
