+++ src/pages/PeoplePage/PeoplePage.tsx
@@ -4,9 +4,11 @@ import { PeopleTable } from './components/PeopleTable/PeopleTable';
 import { PeopleHeader } from './components/PeopleHeader/PeopleHeader';
 import { PeopleList } from './components/PeopleList/PeopleList';
 import { PeopleFilterSort } from './components/PeopleFilterSort/PeopleFilterSort';
+import { usePeopleFiltered } from '../../hooks/usePeopleFiltered';
 
 export function PeoplePage () {
     const deviceType = useViewport();
+    const { sortOptions, setSortOptions, filteredPeople, giftStatsByPerson } = usePeopleFiltered();
 
     return (
         <section className={styles.peoplePage}>
@@ -14,12 +16,18 @@ export function PeoplePage () {
             {/* Rename this to PeopleHeader */}
             <PeopleHeader />
 
-            {/* <PeopleFilterSort
-
-            /> */}
+            <PeopleFilterSort
+                sortOptions={sortOptions}
+                setSortOptions={setSortOptions}
+            />
 
             {/* Load table version if user is on desktop sized screen (768+px) */}
-            {deviceType === 'desktop' && ( <PeopleTable /> )}
+            {deviceType === 'desktop' && ( 
+                <PeopleTable 
+                    people={filteredPeople}
+                    giftStatsByPerson={giftStatsByPerson}
+                /> 
+            )}
 
             {/* Load card list version if user is on mobile sized screen (below 768px) */}
             {deviceType === 'mobile' && ( <PeopleList /> )}
