+++ src/pages/PeoplePage/components/PeopleFilterSort/PeopleFilterSort.module.css
@@ -1,0 +1,10 @@
+.sortForm {
+    background-color: white;
+
+    padding: 1rem;
+
+    border-radius: 1rem;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+}
