+++ src/pages/PeoplePage/components/PeopleTable/PeopleTable.module.css
@@ -5,33 +5,11 @@
     border-radius: 1rem;
 
     box-shadow: 0 4px 20px rgba(0, 0, 0, 0.06);
-
-}
-
-.peopleTableHeader {
-    padding: 1rem;
-
-    background: linear-gradient(135deg, #f8fafc 0%, #f1f5f9 100%);
-
-    /* border-radius: inherit; */
-    border-radius: 1rem 1rem 0 0;
-
-    /* display: flex;
-    align-items: center;
-    justify-content: space-between; */
-}
-
-.sortForm {
-    display: flex;
-    align-items: center;
-    gap: 1rem;
 }
 
 .peopleTable {
     width: 100%;
     border-collapse: collapse;
-
-
 }
 
 .tableHeader {
